image_moderation/requirements.txt
streamlit
torch
transformers
langchain_groq
pillow
